+++ src/discoverylab/telebot/master/arms/parser/YEIDataParser.java
@@ -22,9 +22,9 @@ public class YEIDataParser extends CoreDataParser{
 		else
 		{
 			instance.setJointType(tokenizer.nextToken());
-			instance.setX(Long.parseLong(tokenizer.nextToken()));
-			instance.setY(Long.parseLong(tokenizer.nextToken()));
-			instance.setZ(Long.parseLong(tokenizer.nextToken()));
+			instance.setX(Integer.parseInt(tokenizer.nextToken()));
+			instance.setY(Integer.parseInt(tokenizer.nextToken()));
+			instance.setZ(Integer.parseInt(tokenizer.nextToken()));
 		}
 
 		return instance;
